Move types to sdk-bundles-api, remove optional dependencies, move responsibility to download scripts into replay-orchestrator
