+++ packages/api/src/index.ts
@@ -47,3 +47,4 @@ export {
   UrlHistoryEvent,
   WindowData,
 } from "./sdk-bundle-api/sdk-to-bundle/session-data";
+export { Replay } from "./replay/replay.types";
