+++ packages/api/src/replay/replay.types.ts
@@ -1,0 +1,6 @@
+import { Project } from "../project.types";
+
+export interface Replay {
+  id: string;
+  project: Project;
+}
