+++ packages/cli/package.json
@@ -26,10 +26,13 @@
     "test": "jest --passWithNoTests"
   },
   "dependencies": {
+    "@alwaysmeticulous/api": "^2.40.0",
+    "@alwaysmeticulous/sdk-bundles-api": "^2.40.0",
     "@alwaysmeticulous/client": "^2.40.0",
     "@alwaysmeticulous/common": "^2.40.0",
     "@alwaysmeticulous/download-helpers": "^2.40.0",
     "@alwaysmeticulous/replay-orchestrator": "^2.40.0",
+    "@alwaysmeticulous/record": "^2.40.0",
     "@alwaysmeticulous/sentry": "^2.40.0",
     "@sentry/node": "^7.36.0",
     "axios": "^1.2.6",
@@ -38,17 +41,8 @@
     "yargs": "^17.5.1"
   },
   "devDependencies": {
-    "@alwaysmeticulous/api": "^2.40.0",
     "@types/yargs": "^17.0.10",
     "@types/luxon": "^3.2.0"
-  },
-  "peerDependencies": {
-    "@alwaysmeticulous/record": "^2.3.3"
-  },
-  "peerDependenciesMeta": {
-    "@alwaysmeticulous/record": {
-      "optional": true
-    }
   },
   "author": {
     "name": "The Meticulous Team",
