+++ packages/cli/src/commands/replay/replay.command.ts
@@ -1,23 +1,22 @@
-import { ScreenshotDiffResult } from "@alwaysmeticulous/api";
+import {
+  ScreenshotDiffResult,
+  Replay,
+  ScreenshotDiffOptions,
+  ScreenshotAssertionsOptions,
+} from "@alwaysmeticulous/api";
+import { performReplay } from "@alwaysmeticulous/replay-orchestrator";
 import {
   GeneratedBy,
-  Replay,
   ReplayExecutionOptions,
   ReplayTarget,
   StoryboardOptions,
-} from "@alwaysmeticulous/common";
-import { loadReplayEventsDependencies } from "@alwaysmeticulous/download-helpers/dist/scripts/replay-assets";
-import { performReplay } from "@alwaysmeticulous/replay-orchestrator";
+} from "@alwaysmeticulous/sdk-bundles-api";
 import { buildCommand } from "../../command-utils/command-builder";
 import {
   COMMON_REPLAY_OPTIONS,
   OPTIONS,
   SCREENSHOT_DIFF_OPTIONS,
 } from "../../command-utils/common-options";
-import {
-  ScreenshotAssertionsOptions,
-  ScreenshotDiffOptions,
-} from "../../command-utils/common-types";
 
 export interface ReplayResult {
   replay: Replay;
@@ -100,7 +99,6 @@ export const rawReplayCommandHandler = async ({
       }
     : { enabled: false };
 
-  const replayEventsDependencies = await loadReplayEventsDependencies();
   const { replay } = await performReplay({
     replayTarget: getReplayTarget({
       appUrl: appUrl ?? null,
@@ -115,7 +113,6 @@ export const rawReplayCommandHandler = async ({
     generatedBy: generatedByOption,
     baseTestRunId: null,
     testRunId: null,
-    replayEventsDependencies,
     debugger: enableStepThroughDebugger,
     suppressScreenshotDiffLogging: false,
   });
