+++ packages/download-helpers/src/scripts/replay-assets.ts
@@ -3,7 +3,6 @@ import { basename, join } from "path";
 import {
   getMeticulousLocalDataDir,
   METICULOUS_LOGGER_NAME,
-  ReplayEventsDependencies,
 } from "@alwaysmeticulous/common";
 import axios from "axios";
 import log from "loglevel";
@@ -21,55 +20,6 @@ interface AssetMetadata {
 
 const ASSETS_FOLDER_NAME = "assets";
 const ASSET_METADATA_FILE_NAME = "assets.json";
-
-export const loadReplayEventsDependencies =
-  async (): Promise<ReplayEventsDependencies> => {
-    const browserUserInteractions = await fetchAsset(
-      "replay/v2/snippet-user-interactions.bundle.js"
-    );
-    const browserPlayback = await fetchAsset(
-      "replay/v2/snippet-playback.bundle.js"
-    );
-    const browserUrlObserver = await fetchAsset(
-      "replay/v2/snippet-url-observer.bundle.js"
-    );
-    const nodeBrowserContext = await fetchAsset(
-      "replay/v2/node-browser-context.bundle.js"
-    );
-    const nodeNetworkStubbing = await fetchAsset(
-      "replay/v2/node-network-stubbing.bundle.js"
-    );
-    const nodeUserInteractions = await fetchAsset(
-      "replay/v2/node-user-interactions.bundle.js"
-    );
-
-    return {
-      browserUserInteractions: {
-        key: "browserUserInteractions",
-        location: browserUserInteractions,
-      },
-      browserPlayback: {
-        key: "browserPlayback",
-        location: browserPlayback,
-      },
-      browserUrlObserver: {
-        key: "browserUrlObserver",
-        location: browserUrlObserver,
-      },
-      nodeBrowserContext: {
-        key: "nodeBrowserContext",
-        location: nodeBrowserContext,
-      },
-      nodeNetworkStubbing: {
-        key: "nodeNetworkStubbing",
-        location: nodeNetworkStubbing,
-      },
-      nodeUserInteractions: {
-        key: "nodeUserInteractions",
-        location: nodeUserInteractions,
-      },
-    };
-  };
 
 export const fetchAsset: (path: string) => Promise<string> = async (path) => {
   const logger = log.getLogger(METICULOUS_LOGGER_NAME);
