+++ packages/replay-orchestrator/package.json
@@ -22,7 +22,9 @@
     "@alwaysmeticulous/common": "^2.40.0",
     "@alwaysmeticulous/client": "^2.40.0",
     "@alwaysmeticulous/download-helpers": "^2.40.0",
+    "@alwaysmeticulous/sdk-bundles-api": "^2.40.0",
     "@alwaysmeticulous/sentry": "^2.40.0",
+    "@alwaysmeticulous/replayer": "^2.40.0",
     "@sentry/node": "^7.36.0",
     "find-free-port": "^2.0.0",
     "adm-zip": "^0.5.9",
@@ -44,14 +46,6 @@
     "@types/pixelmatch": "^5.2.4",
     "@types/pngjs": "^6.0.1",
     "@types/adm-zip": "^0.5.0"
-  },
-  "peerDependencies": {
-    "@alwaysmeticulous/replayer": "^2.5.1"
-  },
-  "peerDependenciesMeta": {
-    "@alwaysmeticulous/replayer": {
-      "optional": true
-    }
   },
   "author": {
     "name": "The Meticulous Team",
