+++ packages/replay-orchestrator/src/index.ts
@@ -1,2 +1,2 @@
-export { performReplay, ReplayOptions } from "./replay/perform-replay";
+export { performReplay } from "./replay/perform-replay";
 export { runAllTests } from "./test-run/parallel-tests/run-all-tests";
