--- packages/replay-orchestrator/src/test-run/parallel-tests/run-all-tests.types.ts
@@ -1,6 +1,0 @@
-export interface TestRunProgress {
-  failedTestCases: number;
-  flakedTestCases: number;
-  passedTestCases: number;
-  runningTestCases: number;
-}
