+++ packages/replay-orchestrator/src/test-run/parallel-tests/utils/config.types.ts
@@ -1,13 +1,5 @@
-import {
-  ScreenshotDiffResult,
-  TestCase,
-  TestCaseResult,
-} from "@alwaysmeticulous/api";
+import { TestCase } from "@alwaysmeticulous/api";
 
 export interface MeticulousCliConfig {
   testCases?: TestCase[];
-}
-
-export interface DetailedTestCaseResult extends TestCaseResult {
-  screenshotDiffResultsByBaseReplayId: Record<string, ScreenshotDiffResult[]>;
 }
