+++ packages/replay-orchestrator/src/test-run/parallel-tests/utils/get-replay-target-for-test-case.ts
@@ -1,5 +1,5 @@
 import { TestCase } from "@alwaysmeticulous/api";
-import { ReplayTarget } from "@alwaysmeticulous/common/dist/types/replay.types";
+import { ReplayTarget } from "@alwaysmeticulous/sdk-bundles-api";
 
 export const getReplayTargetForTestCase = ({
   appUrl,
