+++ packages/replayer/src/index.ts
@@ -1,7 +1,8 @@
-export { replayEvents } from "./replayer";
+export { replayEvents, ReplayEventsOptions } from "./replayer";
 export {
   getStartUrl,
   getOriginalSessionStartUrl,
   getSessionStartTime,
   patchDate,
 } from "./replay.utils";
+export { ReplayEventsDependencies } from "./replay.types";
