+++ packages/replayer/src/replay.types.ts
@@ -1,4 +1,5 @@
-import { RecordedSession, GeneratedBy } from "@alwaysmeticulous/common";
+import { RecordedSession } from "@alwaysmeticulous/common";
+import { GeneratedBy } from "@alwaysmeticulous/sdk-bundles-api";
 import { ConsoleMessageLocation, ConsoleMessageType } from "puppeteer";
 import { AssetSnapshotsData } from "./assets/assets.types";
 
@@ -33,4 +34,18 @@ export interface ReplayData {
     console: ConsoleMessage[];
     pageErrors: PageError[];
   };
+}
+
+export interface ReplayEventsDependency<Key extends string> {
+  key: Key;
+  location: string;
+}
+
+export interface ReplayEventsDependencies {
+  browserUserInteractions: ReplayEventsDependency<"browserUserInteractions">;
+  browserPlayback: ReplayEventsDependency<"browserPlayback">;
+  browserUrlObserver: ReplayEventsDependency<"browserUrlObserver">;
+  nodeBrowserContext: ReplayEventsDependency<"nodeBrowserContext">;
+  nodeNetworkStubbing: ReplayEventsDependency<"nodeNetworkStubbing">;
+  nodeUserInteractions: ReplayEventsDependency<"nodeUserInteractions">;
 }
