+++ packages/sdk-bundles-api/src/replay-orchestrator/bundle-to-sdk/execute-replay.ts
@@ -1,0 +1,10 @@
+import { Replay, ScreenshotDiffResult } from "@alwaysmeticulous/api";
+
+export interface ReplayResult {
+  replay: Replay;
+
+  /**
+   * Empty if screenshottingOptions.enabled was false.
+   */
+  screenshotDiffResultsByBaseReplayId: Record<string, ScreenshotDiffResult[]>;
+}
